+++ test/core.jl
@@ -7698,3 +7698,23 @@ end
     @test a == 1
     @test b == Core.svec(2, 3)
 end
+
+@testset "setproperty! on modules" begin
+    m = Module()
+    @eval m global x::Int
+
+    Core.setglobal!(m, :x, 1)
+    @test m.x === 1
+    Core.setglobal!(m, :x, 2, :release)
+    @test m.x === 2
+    @test_throws ConcurrencyViolationError Core.setglobal!(m, :x, 3, :not_atomic)
+    @test_throws ErrorException Core.setglobal!(m, :x, 4., :release)
+
+    m.x = 1
+    @test m.x === 1
+    setproperty!(m, :x, 2, :release)
+    @test m.x === 2
+    @test_throws ConcurrencyViolationError setproperty!(m, :x, 3, :not_atomic)
+    m.x = 4.
+    @test m.x === 4
+end